requirements.txt
agno>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
apify-client>=1.0.0
flask>=3.0.0
pypdf>=3.17.0
werkzeug>=3.0.0
reportlab>=4.0.0

